pyproject.toml
[tool.poetry]
name = "expenses-git"
version = "0.1.0"
description = ""
authors = ["To Kirill <tio.kirill@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
gunicorn = "^23.0.0"
django = "^5.1.3"
djangorestframework = "^3.15.2"
markdown = "^3.7"
drf-spectacular = "^0.27.2"
django-debug-toolbar = "^4.4.6"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
